Send entire downloader context to progress routines
