+++ src/bar.h
@@ -28,12 +28,14 @@
 #ifndef _WGET_BAR_H
 # define _WGET_BAR_H
 
+#include "progress.h"
+
 void bar_init(void);
 void bar_deinit(void);
 void bar_print(int slotpos, const char *s) G_GNUC_WGET_NONNULL_ALL;
 void bar_printf(int slotpos, const char *fmt, ...) G_GNUC_WGET_PRINTF_FORMAT(2,3) G_GNUC_WGET_NONNULL_ALL;
 void bar_vprintf(int slotpos, const char *fmt, va_list args) G_GNUC_WGET_PRINTF_FORMAT(2,0) G_GNUC_WGET_NONNULL_ALL;
-void bar_update(int slotpos, off_t max, off_t cur, const char *filename);
+void bar_update(struct _body_callback_context *ctx) G_GNUC_WGET_NONNULL_ALL;
 
 /*
 ssize_t
